:recycle: Refactor code management method
