+++ Widget/VisitBundle.swift
@@ -5,6 +5,7 @@
 //  Created by AkkeyLab on 2023/03/12.
 //
 
+import Activity
 import Widget
 import WidgetKit
 import SwiftUI
@@ -23,5 +24,27 @@ struct VisitPreviews: PreviewProvider {
             .previewContext(WidgetPreviewContext(family: .systemMedium))
         VisitView.mock
             .previewContext(WidgetPreviewContext(family: .accessoryInline))
+    }
+}
+
+@available(iOS 16.2, *)
+struct VisitLiveActivityPreviews: PreviewProvider {
+    static let attributes = VisitAttributes(name: "Me")
+    static let contentState = VisitAttributes.ContentState(value: 3)
+
+    @available(iOS 16.2, *)
+    static var previews: some View {
+        attributes
+            .previewContext(contentState, viewKind: .dynamicIsland(.compact))
+            .previewDisplayName("Island Compact")
+        attributes
+            .previewContext(contentState, viewKind: .dynamicIsland(.expanded))
+            .previewDisplayName("Island Expanded")
+        attributes
+            .previewContext(contentState, viewKind: .dynamicIsland(.minimal))
+            .previewDisplayName("Minimal")
+        attributes
+            .previewContext(contentState, viewKind: .content)
+            .previewDisplayName("Notification")
     }
 }
